refactor: separate ai answer types and api (#107)
