+++ src/types/api-response/chatbot.ts
@@ -1,0 +1,7 @@
+export type QnaAiAnswer = {
+  id: number
+  question_id: number
+  output: string
+  using_model: string
+  created_at: string
+}
