+++ src/types/api-response/detail.ts
@@ -31,14 +31,6 @@ export type QnaAnswer = {
   comments: QnaAnswerComment[]
 }
 
-export type QnaAiAnswer = {
-  id: number
-  question_id: number
-  output: string
-  using_model: string
-  created_at: string
-}
-
 export type QnaQuestionDetail = {
   id: number
   title: string
