+++ src/types/index.ts
@@ -1,5 +1,6 @@
 // types
 export * from './api-response/category'
 export * from './api-response/chat'
+export * from './api-response/chatbot'
 export * from './api-response/detail'
 export * from './api-response/qna'
